chore(finance): rename scalar to arithmetic
