+++ platform/packages/finance/src/arithmetic.rs
