+++ platform/packages/finance/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod arithmetic;
 pub mod coin;
 pub mod duration;
 pub mod error;
@@ -10,7 +11,6 @@ pub mod period;
 pub mod price;
 pub mod range;
 pub mod ratio;
-pub mod scalar;
 pub mod zero;
 
 #[cfg(any(test, feature = "testing"))]
